+++ CHANGELOG.md
@@ -1,1 +1,5 @@
 # Change log
+
+## DEV
+
+ * Show the current page title and URL in the preview bar (see contao/core-bundle#1640).